// fct-frontend/src/api/models/appointment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API for the website of \"Academia Fabiana Perez\"
 * REST API for the website of \"Academia Fabiana Perez\"
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AppointmentStatusEnum } from './appointment-status-enum';

/**
 * 
 * @export
 * @interface Appointment
 */
export interface Appointment {
    /**
     * 
     * @type {number}
     * @memberof Appointment
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof Appointment
     */
    'date': string;
    /**
     * 
     * @type {number}
     * @memberof Appointment
     */
    'duration': number;
    /**
     * 
     * @type {AppointmentStatusEnum}
     * @memberof Appointment
     */
    'status': AppointmentStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Appointment
     */
    'description': string;
    /**
     * 
     * @type {number}
     * @memberof Appointment
     */
    'price': number;
    /**
     * 
     * @type {number}
     * @memberof Appointment
     */
    'userId': number;
    /**
     * 
     * @type {number}
     * @memberof Appointment
     */
    'categoryId': number;
}