// fct-frontend/src/api/models/appointment-category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API for the website of \"Academia Fabiana Perez\"
 * REST API for the website of \"Academia Fabiana Perez\"
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AppointmentCategory
 */
export interface AppointmentCategory {
    /**
     * 
     * @type {number}
     * @memberof AppointmentCategory
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof AppointmentCategory
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof AppointmentCategory
     */
    'quotePerHour': number;
}